Implement security measures
In order to perform any operations client is required to provide username and authorization token. Except for creating a user and accessing /login. Once a client has been validated at /login a token will be provided by the server and will be used by the client to inform the server that he is who he claimed to be.

Should the token does not match with the username the client will not be able to perform any logged in operations.

Fixes gh-12

See gh-11