Fix broken javascript unit test
Javascript unit test was unable to execute due to missing modules. By adding karma-cli, the command karma can be executed.  Another fix is adding dependency to angular-resource and incorrect path to the project's js files. Also replace usage of  chrome and firefox launcher with phantomJS as this test will be executed in a headless environment.

gh-2