Replace bootstrap dropdown with a normal link
Angular bootstrap causes build failure as we are not using a suitable
Angular version. However to degrade the version only for one of
angular-bootstrap's functionality was not worth an effort.

As a solution we removed all dependencies on angular-bootstrap and
bootstrap. And replace it with materia design icon (mdi) and a normal
link.

In future perhaps we will use angular-material's menu.

Fixes gh-7 and gh-8.